Fix: ProgTimer is larger than 8 bit
Change ProgTimer from uint_fast8_t to uint32_t. (uint_fast8_t was
already a 32 bit type, so this is a formal change only)

+++ src/HomieDoorOpener.h
@@ -43,7 +43,7 @@ private:
 
     Atm_timer timer_buz;
 
-    static const uint_fast8_t ProgTimer = 10000;
+    static const uint32_t ProgTimer = 10000;
     Atm_timer timer_prog;
